update name already exist for manage and modify session

+++ admin/reportDelete.php
@@ -1,20 +1,16 @@
 <?php
-
 session_start();
 include '../function.php';
 $id = $_GET['delete'];
 $data = new Common();
-// if($_SESSION['id'] = $id) {
-//     $sqlde = $data->getReturnData("DELETE FROM report WHERE id = ? ", [$id]);
-//     header("location: index.php");
-// }
-// else {
-//     header("location: ../login.php");
-// }
-
-$sqlde = $data->getReturnData("DELETE FROM report WHERE id = ? ", [$id]);
-// $sqlde = $data->getReturnData("DELETE FROM report WHERE id = ? , adminId = ?", [$id, $_SESSION['id']]);
-header("location: index.php");
+$row = $data->getOneRowData("SELECT * FROM report WHERE id = ? ", [$id]);
+if($_SESSION['id'] == $row['adminId']) { 
+    $sqlde = $data->getReturnData("DELETE FROM report WHERE id = ? ", [$id]);
+    header("location: index.php");
+}
+else{
+    echo "This page will allow by user only. Please login";
+}
 
 
 ?>
